Backend/requirements.txt
python-dotenv
yt-dlp
dotenv
requests
ipykernel
google-generativeai
fastapi
uvicorn
pydantic